Liquibase add

+++ src/main/resources/liquibase/scripts/script.sql
@@ -2,8 +2,8 @@
 
 -- changeSet see1rg:1
 CREATE TABLE account (
-                         id INT AUTO_INCREMENT,
+                         id BIGINT AUTO_INCREMENT,
                          name VARCHAR(35) NOT NULL,
                          pin VARCHAR(4) NOT NULL,
-                         balance BIGINT
+                         balance NUMERIC
 );
